Move cJSON float helpers to cjson_utils
